@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { ErrorService } from '../../services/error.service';
-import { Validators, FormBuilder, AbstractControl } from '@angular/forms';
+import { Validators, UntypedFormBuilder, AbstractControl } from '@angular/forms';
 import { LoginResult, UserType } from '../../model/generated-models';
 import { SignupInfoPanelComponent } from './signup-info.component';
 import { Subscription } from 'rxjs';
@@ -103,7 +103,7 @@ export class SignUpPanelComponent implements OnInit, OnDestroy {
     constructor(
         private auth: AuthService,
         private errorHandler: ErrorService,
-        private formBuilder: FormBuilder) {
+        private formBuilder: UntypedFormBuilder) {
         this.termsLink = EnvService.terms_link;
         this.privacyLink = EnvService.privacy_link;
         this.showPrivacyLink = EnvService.show_privacy_link;
