@@ -1,5 +1,5 @@
 import { AfterViewInit, Component, ElementRef, EventEmitter, Host, Input, Optional, Output, SkipSelf, ViewChild } from '@angular/core';
-import { ControlContainer, ControlValueAccessor, FormControl, FormControlDirective, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { ControlContainer, ControlValueAccessor, UntypedFormControl, FormControlDirective, NG_VALUE_ACCESSOR } from '@angular/forms';
 
 @Component({
     selector: 'app-form-editbox',
@@ -16,13 +16,13 @@ export class FormEditBoxComponent implements ControlValueAccessor, AfterViewInit
     @ViewChild(FormControlDirective, { static: true })
     formControlDirective!: FormControlDirective;
     @Input() maxlength = 0;
-    @Input() formControl!: FormControl;
+    @Input() formControl!: UntypedFormControl;
     @Input() formControlName!: string;
     @Output() onSubmit = new EventEmitter();
 
     active = true;
 
-    get control(): FormControl {
+    get control(): UntypedFormControl {
         return this.formControl || this.controlContainer.control?.get(this.formControlName);
     }
 
