@@ -1,5 +1,5 @@
 import { Component, Host, Input, OnDestroy, OnInit, Optional, SkipSelf, ViewChild } from '@angular/core';
-import { AbstractControl, ControlContainer, ControlValueAccessor, FormControl, FormControlDirective, NG_VALUE_ACCESSOR, Validators } from '@angular/forms';
+import { AbstractControl, ControlContainer, ControlValueAccessor, UntypedFormControl, FormControlDirective, NG_VALUE_ACCESSOR, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 
 @Component({
@@ -24,7 +24,7 @@ export class FormTextBoxComponent implements ControlValueAccessor, OnInit, OnDes
     @Input() inputMask = '';
     @Input() maxlength = 0;
     @Input() errorMessages: { [key: string]: string } = {};
-    @Input() formControl!: FormControl;
+    @Input() formControl!: UntypedFormControl;
     @Input() formControlName!: string;
     @Input() numberField = false;
     @Input() inputMode = 'text';
@@ -39,7 +39,7 @@ export class FormTextBoxComponent implements ControlValueAccessor, OnInit, OnDes
     errorMessage = '';
     requiredFlag = false;
 
-    get control(): FormControl {
+    get control(): UntypedFormControl {
         this.requiredFlag = false;
         const c = this.formControl || this.controlContainer.control?.get(this.formControlName);
         if (c && c.validator) {
