@@ -1,5 +1,5 @@
 import { Component, Inject } from '@angular/core';
-import { AbstractControl, FormBuilder, Validators } from '@angular/forms';
+import { AbstractControl, UntypedFormBuilder, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
 import { DialogData } from 'src/app/model/dialog.model';
 
@@ -22,7 +22,7 @@ export class TwoFaDialogBox {
     }
 
     constructor(
-        private formBuilder: FormBuilder,
+        private formBuilder: UntypedFormBuilder,
         public dialogRef: MatDialogRef<TwoFaDialogBox>,
         @Inject(MAT_DIALOG_DATA)
         public data: DialogData) { }
