@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { AbstractControl, FormBuilder, Validators } from '@angular/forms';
+import { AbstractControl, UntypedFormBuilder, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 import { ContactsFilter, FilterChip, FilterChipType, ProfileBaseFilter, TransactionsFilter } from 'src/app/model/filter.model';
 import { CurrencyView } from 'src/app/model/payment.model';
@@ -49,7 +49,7 @@ export class ContactsFilterBarComponent implements OnInit, OnDestroy {
         return this.filterForm.get('zeroBalance');
     }
 
-    constructor(private formBuilder: FormBuilder) { }
+    constructor(private formBuilder: UntypedFormBuilder) { }
 
     ngOnInit(): void {
         this.subscriptions.add(
