@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, OnDestroy, Output } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { ErrorService } from '../../services/error.service';
-import { Validators, FormBuilder, AbstractControl } from '@angular/forms';
+import { Validators, UntypedFormBuilder, AbstractControl } from '@angular/forms';
 import { LoginResult, PostAddress, SettingsKyc, UserType } from '../../model/generated-models';
 import { Subscription } from 'rxjs';
 import { getCountryByCode3 } from '../../model/country-code.model';
@@ -43,7 +43,7 @@ export class CompanyLevelVerificationComponent implements OnDestroy{
     constructor(
         private auth: AuthService,
         private errorHandler: ErrorService,
-        private formBuilder: FormBuilder) {
+        private formBuilder: UntypedFormBuilder) {
         this.companyNameControl = this.infoForm.get('companyName');
     }
 
